[AIP-21] Move contrib.hooks.vertica_hook providers.vertica.hooks.vertica
