+++ airflow/contrib/operators/vertica_to_hive.py
@@ -26,9 +26,9 @@ from tempfile import NamedTemporaryFile
 
 import unicodecsv as csv
 
-from airflow.contrib.hooks.vertica_hook import VerticaHook
 from airflow.models import BaseOperator
 from airflow.providers.apache.hive.hooks.hive import HiveCliHook
+from airflow.providers.vertica.hooks.vertica import VerticaHook
 from airflow.utils.decorators import apply_defaults
