+++ airflow/contrib/operators/vertica_to_mysql.py
@@ -23,9 +23,9 @@ from tempfile import NamedTemporaryFile
 import MySQLdb
 import unicodecsv as csv
 
-from airflow.contrib.hooks.vertica_hook import VerticaHook
 from airflow.models import BaseOperator
 from airflow.providers.mysql.hooks.mysql import MySqlHook
+from airflow.providers.vertica.hooks.vertica import VerticaHook
 from airflow.utils.decorators import apply_defaults
