+++ airflow/providers/vertica/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.vertica.hooks.vertica`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.vertica.hooks.vertica import VerticaHook  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.vertica.hooks.vertica`.",
-    DeprecationWarning, stacklevel=2
-)
